# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CCE API",
    "description": "Code context engine: indexing, search, relations and tool endpoints",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/tools/definition": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Handle goto definition",
        "description": "# Endpoint\n\n`POST /api/tools/definition`",
        "operationId": "handle_goto_definition",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GotoDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Definition result, errors reported in-band",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GotoDefinitionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tools/keyword-search": {
      "post": {
        "tags": [
          "Tools"
        ],
        "operationId": "handle_keyword_search",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeywordSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Keyword search result, errors reported in-band",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeywordSearchApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/references": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Handle find references",
        "description": "# Endpoint\n\n`POST /api/tools/references`",
        "operationId": "handle_find_references",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindReferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "References result, errors reported in-band",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindReferencesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/symbols": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Handle get symbols",
        "description": "# Endpoint\n\n`POST /api/tools/symbols`",
        "operationId": "handle_get_symbols",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetSymbolsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Symbols result, errors reported in-band",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSymbolsResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DefinitionCode": {
        "type": "object",
        "description": "Definition code with metadata",
        "required": [
          "location",
          "name",
          "kind",
          "code",
          "signature"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Definition code (signature only or full body)"
          },
          "kind": {
            "type": "string",
            "description": "Symbol kind (LSP SymbolKind name)"
          },
          "location": {
            "$ref": "#/components/schemas/DefinitionLocation",
            "description": "Definition location"
          },
          "name": {
            "type": "string",
            "description": "Symbol name"
          },
          "signature": {
            "type": "string",
            "description": "Signature"
          }
        }
      },
      "DefinitionLocation": {
        "type": "object",
        "description": "Definition location",
        "required": [
          "path",
          "entity_id",
          "line",
          "end_line"
        ],
        "properties": {
          "end_line": {
            "type": "integer",
            "description": "End line number (1-based)",
            "minimum": 0
          },
          "entity_id": {
            "type": "integer",
            "format": "int64",
            "description": "Entity ID",
            "minimum": 0
          },
          "line": {
            "type": "integer",
            "description": "Start line number (1-based)",
            "minimum": 0
          },
          "path": {
            "type": "string",
            "description": "File path"
          }
        }
      },
...
      "FileSymbolResult": {
        "type": "object",
        "description": "Result for a single file's symbols",
        "required": [
          "path",
          "success"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Error message (if failed)"
          },
          "path": {
            "type": "string",
            "description": "File path"
          },
          "success": {
            "type": "boolean",
            "description": "Whether the operation succeeded"
          },
          "symbol_count": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Number of symbols (if successful)",
            "minimum": 0
          },
          "symbols": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/SymbolInfo"
            },
            "description": "Symbol list (if successful)"
          }
        }
      },
      "FindReferencesRequest": {
        "type": "object",
        "description": "Find references request",
        "required": [
          "project_id",
          "path",
          "line"
        ],
        "properties": {
          "column": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "context_lines": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "include_entity_info": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "include_snippet": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "line": {
            "type": "integer",
            "minimum": 0
          },
          "path": {
            "type": "string"
          },
          "project_id": {
            "type": "integer",
            "format": "int64"
          },
          "symbol": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "FindReferencesResponse": {
        "type": "object",
        "description": "Find references API response",
        "required": [
          "success"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "relation_info": {
            "type": [
              "object",
              "null"
            ],
            "description": "Relation capability state when the index is degraded"
          },
          "result": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FindReferencesResult"
              }
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "FindReferencesResult": {
        "type": "object",
        "description": "Find references result payload",
        "required": [
          "total_count",
          "file_count",
          "references"
        ],
        "properties": {
          "file_count": {
            "type": "integer",
            "description": "Number of files containing references",
            "minimum": 0
          },
          "references": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupedReferences"
            },
            "description": "References grouped by file"
          },
          "symbol": {
            "type": [
              "string",
              "null"
            ],
            "description": "Symbol name (if provided)"
          },
          "total_count": {
            "type": "integer",
            "description": "Total number of references",
            "minimum": 0
          }
        }
      },
...
      "GetSymbolsRequest": {
        "type": "object",
        "description": "Get symbols request",
        "required": [
          "project_id",
          "paths"
        ],
        "properties": {
          "paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "project_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "GetSymbolsResponse": {
        "type": "object",
        "description": "Get symbols API response",
        "required": [
          "success"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "relation_info": {
            "type": [
              "object",
              "null"
            ],
            "description": "Relation capability state when the index is degraded"
          },
          "result": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GetSymbolsResult"
              }
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "GetSymbolsResult": {
        "type": "object",
        "description": "Get symbols result payload",
        "required": [
          "results",
          "success_count",
          "fail_count"
        ],
        "properties": {
          "fail_count": {
            "type": "integer",
            "description": "Number of failed operations",
            "minimum": 0
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileSymbolResult"
            },
            "description": "Results for each file"
          },
          "success_count": {
            "type": "integer",
            "description": "Number of successful operations",
            "minimum": 0
          }
        }
      },
      "GotoDefinitionRequest": {
        "type": "object",
        "description": "Goto definition request",
        "required": [
          "project_id",
          "path",
          "line"
        ],
        "properties": {
          "column": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "include_body": {
            "type": "boolean"
          },
          "line": {
            "type": "integer",
            "minimum": 0
          },
          "path": {
            "type": "string"
          },
          "project_id": {
            "type": "integer",
            "format": "int64"
          },
          "symbol": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "GotoDefinitionResponse": {
        "type": "object",
        "description": "Goto definition API response",
        "required": [
          "success"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "relation_info": {
            "type": [
              "object",
              "null"
            ],
            "description": "Relation capability state when the index is degraded"
          },
          "result": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GotoDefinitionResult"
              }
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "GotoDefinitionResult": {
        "type": "object",
        "description": "Goto definition result payload",
        "required": [
          "definitions"
        ],
        "properties": {
          "definitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DefinitionCode"
            },
            "description": "List of definitions (may be multiple for interface implementations)"
          },
          "symbol": {
            "type": [
              "string",
              "null"
            ],
            "description": "Symbol name (if provided)"
          }
        }
      },
...
      "GroupedReferences": {
        "type": "object",
        "description": "References grouped by file",
        "required": [
          "path",
          "count",
          "references"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "description": "Number of references in this file",
            "minimum": 0
          },
          "path": {
            "type": "string",
            "description": "File path"
          },
          "references": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReferenceLocation"
            },
            "description": "List of references"
          }
        }
      },
...
      "KeywordSearchApiResponse": {
        "type": "object",
        "description": "Keyword search API response",
        "required": [
          "success"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "result": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/KeywordSearchResult"
              }
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "KeywordSearchItem": {
        "type": "object",
        "description": "A single keyword search result with highlighted snippet",
        "required": [
          "chunk_id",
          "score",
          "file_path",
          "title",
          "highlighted_snippet",
          "start_line",
          "end_line"
        ],
        "properties": {
          "chunk_id": {
            "type": "string",
            "description": "Chunk/document ID"
          },
          "end_line": {
            "type": "integer",
            "format": "int32",
            "description": "End line in the file",
            "minimum": 0
          },
          "file_path": {
            "type": "string",
            "description": "File path containing the match"
          },
          "highlighted_snippet": {
            "type": "string",
            "description": "Highlighted code snippet (HTML with <mark> tags)"
          },
          "score": {
            "type": "number",
            "format": "float",
            "description": "BM25 relevance score"
          },
          "start_line": {
            "type": "integer",
            "format": "int32",
            "description": "Start line in the file",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Entity/function title"
          }
        }
      },
      "KeywordSearchRequest": {
        "type": "object",
        "description": "Keyword search request",
        "required": [
          "query",
          "top_n",
          "project_id"
        ],
        "properties": {
          "epoch": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Optional epoch for version-aware filtering"
          },
          "project_id": {
            "type": "integer",
            "format": "int64",
            "description": "Project ID for scoped search"
          },
          "query": {
            "type": "string",
            "description": "Search query text (must be non-empty)"
          },
          "term_operator": {
            "$ref": "#/components/schemas/KeywordTermOperator",
            "description": "Operator for combining multiple query terms"
          },
          "top_n": {
            "type": "integer",
            "description": "Maximum number of results to return (must be > 0)",
            "minimum": 0
          }
        }
      },
      "KeywordSearchResult": {
        "type": "object",
        "description": "Keyword search result payload",
        "required": [
          "query",
          "total",
          "results"
        ],
        "properties": {
          "query": {
            "type": "string",
            "description": "The original query"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KeywordSearchItem"
            },
            "description": "Search results with highlighted snippets"
          },
          "total": {
            "type": "integer",
            "description": "Total number of results returned",
            "minimum": 0
          }
        }
      },
      "KeywordTermOperator": {
        "type": "string",
        "description": "Operator for combining multiple query terms",
        "enum": [
          "or",
          "and"
        ]
      },
...
      "SymbolInfo": {
        "type": "object",
        "description": "Symbol information",
        "required": [
          "name",
          "kind",
          "line",
          "end_line"
        ],
        "properties": {
          "children": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/SymbolInfo"
            },
            "description": "Child symbols"
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "Detail information (signature)"
          },
          "end_line": {
            "type": "integer",
            "description": "End line number (1-based)",
            "minimum": 0
          },
          "kind": {
            "type": "string",
            "description": "Symbol kind (LSP SymbolKind name)"
          },
          "line": {
            "type": "integer",
            "description": "Start line number (1-based)",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "description": "Symbol name"
          }
        }
      },
...
    }
  },
...
}